UPDATE add comments api POST, PATCH marshmallow validate

+++ app/api/board/comment/service.py
@@ -6,7 +6,11 @@ from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt_claims
 from app import db
 
 from app.api.board.post.model import PostModel
-from app.api.board.comment.model import CommentModel, comments_schema, comments_schema_user
+from app.api.board.comment.model import CommentModel,\
+                                        comments_schema,\
+                                        comments_schema_user,\
+                                        CommentInputSchema,\
+                                        CommentPatchInputSchema
 from app.api.user.account.model import AccountModel
 from app.api.user.model import UserModel
 
@@ -17,7 +21,6 @@ def check_user_permission(comment, admin_allow):
     request_user = AccountModel.get_user_by_email(email=get_jwt_identity())
 
     if admin_allow:
-        roles = get_jwt_claims()['roles']
         roles = get_jwt_claims()['roles']
         if roles == 'admin':
             return True
@@ -33,20 +36,24 @@ class CommentService():
     @jwt_required
     def modify_comment(comment_id):
         comment = CommentModel.query.get(comment_id)
+        json = request.get_json()
+
         if not comment:
             return jsonify(msg='wrong comment_id, comment not found'), 404
 
         if not check_user_permission(comment=comment, admin_allow=False):
             return jsonify(msg=f'access denied, u r not {comment.writer.username}'), 403
 
-        new_content = request.json.get('content', None)
-        if not new_content:
-            return jsonify(msg='content parameter missed'), 400
+        CommentInputSchema().validate(json)
 
-        if not is_correct_length(len(new_content)):
-            return jsonify(msg='content is too long, it must be shorter than 100'), 413
+        # new_content = request.json.get('content', None)
+        # if not new_content:
+        #     return jsonify(msg='content parameter missed'), 400
+        #
+        # if not is_correct_length(len(new_content)):
+        #     return jsonify(msg='content is too long, it must be shorter than 100'), 413
 
-        comment.content = new_content
+        comment.content = json['content']
         comment.wrote_datetime = datetime.now()
 
         db.session.commit()
@@ -61,8 +68,8 @@ class CommentService():
         if not comment:
             return jsonify(msg='wrong comment_id, comment not found'), 404
 
-        # if not check_user_permission(comment=comment, admin_allow=True):
-        #     return jsonify(msg=f'access denied, u r not {comment.writer.username}'), 403
+        if not check_user_permission(comment=comment, admin_allow=True):
+            return jsonify(msg=f'access denied, u r not {comment.writer.username}'), 403
 
         comment.delete_comment()
         db.session.commit()
@@ -82,37 +89,35 @@ class CommentListService():
         if not PostModel.query.get(post_id):
             return jsonify(msg='wrong post_id, post not found'), 404
 
+        json = request.get_json()
+        error = CommentInputSchema().validate(json)
+        if error:
+            return jsonify(msg='json validate error'), 400
 
-        content = request.json.get('content', None)
-        if not content:
-            return jsonify(msg='content parameter missed'), 400
-        if not is_correct_length(len(content)):
-            return jsonify(msg='content is too long, it must be shorter than 100'), 413
+        content = json.get('content', None)
+        upper_comment_id = json.get('upper_comment_id')
+
+        if upper_comment_id:
+            upper_comment = CommentModel.query.get(upper_comment_id)
+            if not upper_comment:
+                return jsonify(msg='upper comment is not found'), 404
+
+            if not upper_comment.wrote_post_id == int(post_id):
+                return jsonify(msg='post_id of upper_comment is not same with post_id that received'), 403
+
+        # if not content:
+        #     return jsonify(msg='content parameter missed'), 400
+        # if not is_correct_length(len(content)):
+        #     return jsonify(msg='content is too long, it must be shorter than 100'), 413
 
 
         writer = AccountModel.query.filter_by(email = get_jwt_identity()).first().user
 
         new_comment = CommentModel(content=content,
                                    wrote_datetime=datetime.now(),
-                                   wrote_user_id=writer.id)
+                                   wrote_user_id=writer.id,
+                                   upper_comment_id=upper_comment_id)
         db.session.add(new_comment)
-        db.session.flush()
-
-        upper_comment_id = request.args.get('upper_comment_id', None)
-        if upper_comment_id:
-            upper_comment = CommentModel.query.get(upper_comment_id)
-            if not upper_comment:
-                return jsonify(msg='upper comment is not found'), 404
-            if not upper_comment.wrote_post_id == int(post_id):
-                return jsonify(msg='post_id of upper_comment is not same with post_id that received'), 403
-
-            new_comment.upper_comment_id = upper_comment_id
-        else:
-            print(new_comment.id)
-            new_comment.upper_comment_id = new_comment.id
-        new_comment.wrote_post_id = post_id
-
-
         db.session.commit()
 
         return jsonify(msg='comment successfully wrote'), 200
